read FileDataComplete insted of FileData for GameObjectData extracting; move min/max flightBox calc to core ( reextract maps/mmaps )

+++ src/server/game/Maps/Map.h
@@ -164,8 +164,8 @@ class GridMap
         uint8* m_uint8_V8;
     };
     // Height level data
-    float* _maxHeight;
-    float* _minHeight;
+    int16* _maxHeight;
+    int16* _minHeight;
     float _gridHeight;
     float _gridIntHeightMultiplier;
